Update Artificer Revised links for new optionalFeatureTypes (#52)
* update links, proficiencies, and power fist

Updated the links for spells in the Warsmith and level 6 feature, updated the proficiencies to be links and show hand crossbows and heavy crossbows, updated the link for the homebrew to the latest version, fixed the power fist incorrect prerequisite

* alphabetical order for warsmith upgrades

Changed the Warsmith upgrades to be in alphabetical order

* fix: gadgetsmith upgrades, links

Fixed the Gadgetsmith upgrades not formatting properly and showing the wrong specialization upgrades at the wrong levels. Added lots of links for spells and fixed some broken ones and links for some items, conditions, and skills

* fix: author name

Author name was Kobold Press, fixed to Tobias Beis

* update: war,gadget,potion to 1.5.1

Updating the Warsmith, Gadgetsmith, and Potionsmith to 1.5.1, capitalizing all appropriate attribute names, adding dice rollers where appropriate

* update: golemsmith and infusionsmith

update all subclasses to 1.5.1 and remove runesmith and replace with the golemsmith and infusionsmith.

added links for spells, items, and dice where appropriate

* fix: typo on infusionsmith upgrades

Infusionsmith Upgrades was erroneously labeled Golemsmith upgrades due to a copy-paste that wasn't corrected

* update: moved all upgrades to optional features

Moved all of the upgrades out of the class page and into the optional features page. Makes it much easier to search and find the right upgrade.

* fix: optional feature links

Updated the optional feature links for each subclass to reflect the new optionalFeatureType definition.

Added links to some optional features in the base class page to reflect the fact that it is no longer on the same page (mostly gadgetsmith)