Update teqqqie; Weird Flavor Creatures.json
Fixed some HP values on Deepstar and Giant Deepstar that had been incorrectly calculated, changed the name of the Corpse Flower to Carrion Flower to avoid conflicting with the new official Corpse Flower creature. Removed a feature from Giant Deepstar that it was not intended to have, and appended additional flavor text.