Adds web spell link to "abiding webs" spell desc
I saw the abiding web spell the other day and noticed the description references the web spell, yet it doesn't link to the web spell itself. This change should (theoretically) create a hyperlink to the web spell.

I also don't really do the code thing that much, so I just looked at the files and how they did things, and I -think- this is the correct way to do what I want to do.
Thanks!