Dungeon Dad 5e Monster Statblocks update
This file, simply called Dungeon Dad Statblocks, is a massive update of Cadowtin's Dungeon Dad 5e Monster Statblocks featuring a lot more monster statblocks (about almost all of the statblocks in the descriptions of all of dungeon dad's videos), with a pair of races the shardmind and the lupin. The one i'm most proud of are the kaiju line of monsters, specificly the template found in their info sections.