Add files via upload
The Armiger Class for 5e, by Blake and Kevin (Blurhy). The armiger is a levitating weapon master class, inspired by swordmages and bladedancers of all kinds. If you ever wanted to be Noctis from FFXV, a Sage from FFXIV, Irelia from League of Legends, Gilgamesh and Archer from the FATE series, or Tenten from Naruto, then this is the class for you.