Demi-Dragon Update 4.7
The Demi-Dragon is a class that is built from the ground up to enable playable dragons while respecting the needs of a player character—it is designed to avoid causing undue problems at any table. As an enormously difficult design challenge to pin down, this content exists to provide solid rules for a character concept that official rules neglect.

This patch introduces two new subclasses, as well as a handful of additional changes and additions to older subclasses. The Embodiment of the Packlord has an abundance of utility and makes for a neat option for roleplay... and if you aren't quite sure what to play, or want to try a bit of everything, the Embodiment of the Echo lets you do that. After four years of dedicated work on this content, the Demi-Dragon has been thoroughly tested and patched; go try it for yourself—it's fun!